requirements.txt
streamlit>=1.20
python-dotenv>=0.21
langchain>=0.0.250
langchain-openai>=0.0.15
openai>=0.27
langchain-community